Standardize seeder name
